Added mutex around sscanf to work around arising problems.
When running the processing pipeline with more than one thread data
corruption occured. This is due to sscanf not being thread safe, which
seems library dependent.
The workaround is a mutex being shared between the differen parsers
restricting the access to sscanf section.
As a consequence scaling of the program may be restricted, as most of
the work is done by sscanf and its subcalls.
Currently it runs fine on my dual-core, but may show significant
bottle-neck at the octocores.